requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters
langchain-groq
faiss-cpu
streamlit
pypdf
sentence-transformers
groq
python-dotenv
Selenium